requirements.txt
discord.py
numpy
requests
pandas
